split in utils_tags.py
